Fix pointer truncation issues in 64 bit build
This patch only fixes the critical ones that prevent the client
from launching. There are still major issues when trying to use
a 64 bit assistant.

+++ OrionUO/Managers/ProfessionManager.cpp
@@ -408,7 +408,7 @@ void CProfessionManager::LoadProfessionDescription()
 	if (file.Load(g_App.UOFilesPath("professn.enu")))
 	{
 		char *ptr = (char*)file.Start;
-		char *end = (char*)((uint)file.Start + file.Size);
+		char *end = (char*)((uintptr_t)file.Start + file.Size);
 
 		vector<string> list;
